\name{coef.nlreg}
\alias{coef.nlreg}
\title{
  Use coef() on a `nlreg' object
}
\description{
  This is a method for the function \code{coef()} for objects 
  inheriting from class \code{nlreg}.  See \code{\link{coef}} for the 
  general behavior of this function and for the interpretation of 
  \code{object}.
}
\usage{
\method{coef}{nlreg}(object, \dots)
}
\seealso{
  \code{\link[nlreg]{nlreg.object}}, \code{\link{coef}}
}
\keyword{internal}
\keyword{methods}
\keyword{nonlinear}
