\name{fitted.nlreg}
\alias{fitted.nlreg}
\alias{fitted.values.nlreg}
\title{
  Use fitted() on a `nlreg' object
}
\description{
  This is a method for the function \code{fitted()} for objects 
  inheriting from class \code{nlreg}.  See \code{\link{fitted}} for the 
  general behavior of this function and for the interpretation of 
  \code{object}.
}
\usage{
\method{fitted}{nlreg}(object, \dots)
\method{fitted.values}{nlreg}(object, \dots)
}
\seealso{
  \code{\link[nlreg]{nlreg.object}}, \code{\link{fitted}}
}
\keyword{internal}
\keyword{methods}
\keyword{nonlinear}
