\name{param.nlreg}
\alias{param.nlreg}
\title{
  Use param() on a `nlreg' object
}
\description{
  This is a method for the function \code{param()} for objects 
  inheriting from class \code{nlreg}.  See \code{\link[nlreg]{param}} 
  for the general behavior of this function and for the interpretation
  of \code{object}.
}
\usage{
\method{param}{nlreg}(object, \dots)
}
\seealso{
  \code{\link[nlreg]{nlreg.object}}, \code{\link[nlreg]{param}}
}
\keyword{methods}
\keyword{regression}
\keyword{nonlinear}
\keyword{internal}
